Added a feature to the SignUp unit test which deletes the created user after each time the test is ran, to avoid errors
When the SignUp unit test is run, there is a user created with a certain email. Currently, if the test is rerun, there is an error since there already exists an email within the Users table with same test email. In order to avoid this, we can delete the test email from the DB at the end of every test